Changes long to int for id Profil

+++ src/main/java/com/awesome/thesis/controller/DateiController.java
@@ -57,7 +57,7 @@ public class DateiController {
             Integer id = auth.getPrincipal().getAttribute("id");
             String dateiId = UUID.randomUUID().toString();
             DateiDTO dateiDTO = new DateiDTO(dateiId, infos.getTitle(), infos.getDescription());
-            profilEditor.addDateiOld(id.longValue(), dateiDTO);
+            profilEditor.addDateiOld(id, dateiDTO);
 
             model.addAttribute("dateiInfos", infos);
             model.addAttribute("nachricht", infos.getTitle() + " wurde erfolgreich hochgeladen.");
